+++ Comportamientos/Motor2D/src/skeletonsprite.cpp
@@ -45,7 +45,7 @@ SkeletonSprite::SkeletonSprite(const String& filename) : Sprite(NULL) {
 	for ( uint32 index = 0; index < bone->CountFrames(); index++ ) {
         lastframe = max(lastframe, bone->GetFrame(index)->GetId());
 	}
-	SetFrameRange(0, static_cast<uint16>( lastframe ) );
+	SetFrameRange(0, lastframe);
 
 	// Eliminamos los datos cargados del XML
 	delete data;
